Refactor boundary classes and ensure backward compatibilities
