+++ pyresample/boundary/simple_boundary.py
@@ -1,0 +1,35 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2014-2023 Pyresample developers
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""The deprecated SimpleBoundary class."""
+
+import logging 
+
+logger = logging.getLogger(__name__)
+
+
+class SimpleBoundary(object):
+    """Container for geometry boundary.
+
+    Labelling starts in upper left corner and proceeds clockwise
+    """
+
+    def __init__(self, side1, side2, side3, side4):
+        self.side1 = side1
+        self.side2 = side2
+        self.side3 = side3
+        self.side4 = side4
